Second maximum function is placed in a separate translation unit
