+++ nikitov.dmitriy/P1/ismax.cpp
@@ -1,0 +1,9 @@
+void isMax(int now, int& actual, int& pre)
+{
+  if (now >= actual){
+    pre = actual;
+    actual = now;
+  } else if (now > pre){
+    pre = now;
+  }
+}
